Alloca to _malloca , memcpy to std, projects update

+++ golded3/gemenu.cpp
@@ -877,7 +877,7 @@ int ChangeUsername()
 
 //  ------------------------------------------------------------------
 
-int ChangeTemplate()
+size_t ChangeTemplate()
 {
     if (not CFG->tpl.empty())
     {
